+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All releases with the relative changes are documented in this file.
 
+## [Unreleased]
+### Added
+- Added `state` field to `Listener` struct, exposing the socket's TCP connection state ([#49](https://github.com/GyulyVGC/listeners/pull/49))
+- New `SocketState` enum ([#49](https://github.com/GyulyVGC/listeners/pull/49))
+
 ## [0.5.1] - 2026-03-02
 ### Fixed
 - Include build.rs in the published crate